rename dir: vue-data-entityに from vueDataEntiry
