+++ resources/ts/components/pages/SearchTeam.vue
@@ -42,8 +42,8 @@
 </template>
 
 <script lang="ts">
-import { FileDataObject } from "../../vueDataEntity/TeamData";
-import { OrderTextValueObject } from "../../vueDataEntity/OrderTextValueObject";
+import { FileDataObject } from "../../vue-data-entity/TeamData";
+import { OrderTextValueObject } from "../../vue-data-entity/OrderTextValueObject";
 
 import { Vue, Component } from "vue-property-decorator";
