+++ resources/ts/vue-data-entity/OrderTextValueObject.ts
