+++ resources/ts/vue-data-entity/TeamData.ts
